run pytests on cpu
