+++ models/src/anemoi/models/distributed/transformer.py
@@ -17,42 +17,7 @@ from torch.distributed.distributed_c10d import ProcessGroup
 
 from anemoi.models.distributed.balanced_partition import get_balanced_partition_sizes
 from anemoi.models.distributed.utils import get_memory_format
-
-
-def _alltoallwrapper(output_list: list, input_list: list, group: ProcessGroup):
-    """Wrapper function for all_to_all across NCCL, MPI and Gloo backends.
-    There is no all_to_all primitive for the Gloo backend. In that case each
-    process broadcasts its tensor asynchronously.
-
-    Retuns nothing but modifies output_list in-place
-
-    """
-    comm_size = dist.get_world_size(group=group)
-
-    if dist.get_backend(group) == "gloo":
-
-        # Need to check torch version here bc the syntax for dist.send/recv changed in torch v2.6
-        torch_version = torch.__version__.split(".")
-        torch_major_version = int(torch_version[0])
-        torch_minor_version = int(torch_version[1])
-        if torch_major_version <= 2 and torch_minor_version < 6:
-            raise NotImplementedError("Gloo all_to_all not implemented for torch < v2.6")
-
-        reqs = []
-        rank = dist.get_rank(group=group)
-        # Here we implement the linear shift algorithm from Hofmann and Ruenger, 2013
-        for i in range(0, comm_size):
-            j = (i - rank + comm_size) % comm_size
-            if j != rank:
-                # exchange data with rank j
-                reqs.append(dist.isend(input_list[j], group_dst=j, group=group))
-                reqs.append(dist.irecv(output_list[j], group_src=j, group=group))
-            else:
-                output_list[rank] = input_list[rank]
-        for req in reqs:
-            req.wait()
-    else:
-        dist.all_to_all(output_list, input_list, group=group)
+from anemoi.models.distributed.primitives import _alltoallwrapper
 
 
 def _headsalltoall(input_: Tensor, shapes: list, group: Optional[ProcessGroup] = None) -> Tensor:
