+++ models/tests/distributed/test_distributed_sharding.py
@@ -69,26 +69,22 @@ def _run_spawn_suite(backend: str, suite: str, nproc_per_node: int = 2) -> None:
             f"stderr:\n{completed.stderr}",
         )
 
-
 @pytest.mark.multigpu
 def test_distributed_graph_primitives_core_nccl() -> None:
+    backend = 'nccl' if torch.cuda.is_available() else 'gloo'
     world_size = _requested_world_size()
-    if not torch.cuda.is_available() or torch.cuda.device_count() < world_size:
-        pytest.skip(f"Requires at least {world_size} CUDA devices for NCCL sharding primitives.")
-    _run_spawn_suite(backend="nccl", suite="core", nproc_per_node=world_size)
+    _run_spawn_suite(backend=backend, suite="core", nproc_per_node=world_size)
 
 
 @pytest.mark.multigpu
 def test_distributed_graph_primitives_channels_nccl() -> None:
+    backend = 'nccl' if torch.cuda.is_available() else 'gloo'
     world_size = _requested_world_size()
-    if not torch.cuda.is_available() or torch.cuda.device_count() < world_size:
-        pytest.skip(f"Requires at least {world_size} CUDA devices for NCCL all_to_all channel sharding.")
-    _run_spawn_suite(backend="nccl", suite="channels", nproc_per_node=world_size)
+    _run_spawn_suite(backend=backend, suite="channels", nproc_per_node=world_size)
 
 
 @pytest.mark.multigpu
 def test_distributed_transformer_sharding_nccl() -> None:
+    backend = 'nccl' if torch.cuda.is_available() else 'gloo'
     world_size = _requested_world_size()
-    if not torch.cuda.is_available() or torch.cuda.device_count() < world_size:
-        pytest.skip(f"Requires at least {world_size} CUDA devices for NCCL transformer sharding.")
-    _run_spawn_suite(backend="nccl", suite="transformer", nproc_per_node=world_size)
+    _run_spawn_suite(backend=backend, suite="transformer", nproc_per_node=world_size)
